스쿨 컬럼 추가, 서버 데이터 삽입 로직 추가

+++ src/main/java/com/app/recychool/service/ReserveQueryServiceImpl.java
@@ -44,7 +44,7 @@ public class ReserveQueryServiceImpl implements ReserveQueryService {
         int deposit;
         String usageTime;
 
-        Integer maxParkingCapacity = null;
+        Double maxParkingCapacity = null;
 
         if (reserveType == ReserveType.PLACE) {
             price = 0;
